Refactor Code, code thêm data

+++ DATN-SPCT/src/main/java/com/example/datnspct/dto/GioHangCTDTO.java
@@ -4,6 +4,10 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -12,5 +16,5 @@ public class GioHangCTDTO {
     private Integer idGH;
     private Integer idSPCT;
     private Integer soLuong;
-    private Double donGia;
-} 
+    private BigDecimal donGia;
+}
